feat(count): implement item counts for characters and episodes
