+++ app/src/main/java/com/character/mobile/data/models/CharacterResponseDto.kt
@@ -1,0 +1,6 @@
+package com.character.mobile.data.models
+
+data class CharacterResponseDto(
+    val info: InfoDto,
+    val results: List<CharacterDto>
+)
