+++ app/src/main/java/com/character/mobile/data/models/EpisodeDto.kt
@@ -1,0 +1,15 @@
+package com.character.mobile.data.models
+
+import com.google.gson.annotations.SerializedName
+
+
+data class EpisodeDto(
+    val id: Long,
+    val name: String,
+    @SerializedName("air_date")
+    val airDate: String,
+    val episode: String,
+    val characters: List<String>,
+    val url: String,
+    val created: String,
+)
