+++ app/src/main/java/com/character/mobile/data/models/InfoDto.kt
@@ -1,0 +1,8 @@
+package com.character.mobile.data.models
+
+data class InfoDto(
+    val count: Int,
+    val pages: Int,
+    val next: String?,
+    val prev: String?
+)
