+++ app/src/main/java/com/character/mobile/data/models/OriginDto.kt
@@ -1,0 +1,6 @@
+package com.character.mobile.data.models
+
+data class OriginDto(
+    val name: String? = null,
+    val url: String? = null
+)
