+++ app/src/main/java/com/character/mobile/data/network/Apis.kt
@@ -1,0 +1,10 @@
+package com.character.mobile.data.network
+
+object Apis {
+   const val BASEURL=  "https://rickandmortyapi.com/api/"
+    const val CHARACTERS = "character"
+    const val SINGLE_CHARACTER = "character/{id}"
+    const val LOCATIONS = "/location"
+    const val SINGLE_LOCATION = "/location/{id}"
+
+}
