+++ app/src/main/java/com/character/mobile/domain/models/CharacterResponse.kt
@@ -1,0 +1,6 @@
+package com.character.mobile.domain.models
+
+data class CharacterResponse(
+    val info: Info,
+    val results: List<Character>
+)
