+++ app/src/main/java/com/character/mobile/domain/models/Episode.kt
@@ -1,0 +1,15 @@
+package com.character.mobile.domain.models
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Episode(
+    val id: Long,
+    val name: String,
+    val airDate: String,
+    val episode: String,
+    val characters: List<String>,
+    val url: String,
+    val created: String,
+) : Parcelable
