+++ app/src/main/java/com/character/mobile/domain/models/Info.kt
@@ -1,0 +1,8 @@
+package com.character.mobile.domain.models
+
+data class Info(
+    val count: Int,
+    val pages: Int,
+    val next: String?,
+    val prev: String?
+)
