+++ app/src/main/java/com/character/mobile/domain/models/Origin.kt
@@ -1,0 +1,10 @@
+package com.character.mobile.domain.models
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Origin(
+    val name: String? = null,
+    val url: String? = null
+) : Parcelable
