--- app/src/main/java/com/example/paginationandroid/data/models/CharacterResponseDto.kt
@@ -1,6 +1,0 @@
-package com.example.paginationandroid.data.models
-
-data class CharacterResponseDto(
-    val info: InfoDto,
-    val results: List<CharacterDto>
-)
