--- app/src/main/java/com/example/paginationandroid/data/models/InfoDto.kt
@@ -1,8 +1,0 @@
-package com.example.paginationandroid.data.models
-
-data class InfoDto(
-    val count: Int,
-    val pages: Int,
-    val next: String?,
-    val prev: String?
-)
