--- app/src/main/java/com/example/paginationandroid/data/models/OriginDto.kt
@@ -1,6 +1,0 @@
-package com.example.paginationandroid.data.models
-
-data class OriginDto(
-    val name: String? = null,
-    val url: String? = null
-)
