--- app/src/main/java/com/example/paginationandroid/data/network/ApiClient.kt
@@ -1,28 +1,0 @@
-package com.example.paginationandroid.data.network
-
-import com.example.paginationandroid.data.models.EpisodeDto
-import com.example.paginationandroid.data.models.CharacterDto
-import com.example.paginationandroid.data.models.CharacterResponseDto
-import retrofit2.Response
-import retrofit2.http.GET
-import retrofit2.http.Path
-import retrofit2.http.Query
-import retrofit2.http.Url
-
-interface ApiClient {
-    @GET(Apis.CHARACTERS)
-    suspend fun getMovies(@Query("page") query: Int): Response<CharacterResponseDto>
-
-    @GET(Apis.SINGLE_CHARACTER)
-    suspend fun getSingleMovie(@Path("id") id: Int): Response<CharacterDto>
-
-    @GET
-    suspend fun getEpisode(
-        @Url url: String
-    ): Response<EpisodeDto>
-
-    @GET
-    suspend fun getCharacter(
-        @Url url: String
-    ): Response<CharacterDto>
-}
