--- app/src/main/java/com/example/paginationandroid/data/network/Apis.kt
@@ -1,10 +1,0 @@
-package com.example.paginationandroid.data.network
-
-object Apis {
-   const val BASEURL=  "https://rickandmortyapi.com/api/"
-    const val CHARACTERS = "character"
-    const val SINGLE_CHARACTER = "character/{id}"
-    const val LOCATIONS = "/location"
-    const val SINGLE_LOCATION = "/location/{id}"
-
-}
