--- app/src/main/java/com/example/paginationandroid/data/network/NetworkResponse.kt
@@ -1,8 +1,0 @@
-package com.example.paginationandroid.data.network
-
-data class NetworkResponse<T>(
-    val success: Boolean = false,
-    val data: T? = null,
-    val error: String? = null,
-    val extra: String? = null
-)
