--- app/src/main/java/com/example/paginationandroid/domain/models/CharacterResponse.kt
@@ -1,6 +1,0 @@
-package com.example.paginationandroid.domain.models
-
-data class CharacterResponse(
-    val info: Info,
-    val results: List<Character>
-)
