--- app/src/main/java/com/example/paginationandroid/domain/models/Episode.kt
@@ -1,15 +1,0 @@
-package com.example.paginationandroid.domain.models
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Episode(
-    val id: Long,
-    val name: String,
-    val airDate: String,
-    val episode: String,
-    val characters: List<String>,
-    val url: String,
-    val created: String,
-) : Parcelable
