--- app/src/main/java/com/example/paginationandroid/domain/models/Info.kt
@@ -1,8 +1,0 @@
-package com.example.paginationandroid.domain.models
-
-data class Info(
-    val count: Int,
-    val pages: Int,
-    val next: String?,
-    val prev: String?
-)
