--- app/src/main/java/com/example/paginationandroid/domain/models/Origin.kt
@@ -1,10 +1,0 @@
-package com.example.paginationandroid.domain.models
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Origin(
-    val name: String? = null,
-    val url: String? = null
-) : Parcelable
